chore: Language fixes
This started as a small tweak to the logging output to make it less ambiguous.  
Specifically, I've added the *No* that was missing in the `KRB5_ADMINSERVER` check, and replaced *in place* with *instead* as I feel that is more descriptive of what we're doing.

While doing that, I noticed a typo (*Exting* instead of *Exiting*) so I decided to go through all the logging output and make it consistent in working and capitalization.

None of these changes impact any code, logging output only.